==> 00a_data_explore.ipynb <==
import pandas as pd
import numpy as np

train_data = pd.read_csv('../data/download/train.csv')

train_data.head()

train_data.iloc[:,1:].describe().round(4).T

for i in range(1000, 1020):
    print('COMMENT:', train_data.iloc[i,1])
    print('\n')
