==> 00c_text_features.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
from string import punctuation
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import roc_curve, auc

profane = pd.read_csv('../data/download/profane.txt', skiprows=1, header=None)
profane = set(profane[0].values)
models = ['toxic','severe_toxic','obscene','threat','insult','identity_hate']

def tags(x):
    x = x.replace('\n',' ')
    table = str.maketrans('', '', punctuation)
    x = x.lower().translate(table)
    text = nltk.word_tokenize(x)
    text = nltk.pos_tag(text)
    return text

train_data = pd.read_csv('../data/download/train.csv')
train_feats = train_data[['id']]
train_labels = train_data.iloc[:, 2:]
train_data = train_data['comment_text']
train_tags = train_data.map(tags)

test_data = pd.read_csv('../data/download/test.csv')
test_feats = test_data[['id']]
test_data = test_data['comment_text']
test_tags = test_data.map(tags)

def count_profane(string):
    count = 0
    for word in profane:
        if string.lower().find(word) > -1:
            count += 1
    return count / max(len(string.split()),1)

def count_capital(string):
    count = 0
    for char in string:
        if char.isupper():
            count += 1
    return count / max(len(string),1)

def count_punct(string):
    count = 0
    string = string.replace('\n','')
    for char in string:
        if char in punctuation:
            count += 1
    return count

def count_pronoun(text):
    return np.sum([1 if 'PRP' in x[1] else 0 for x in text]) / max(len(text),1)

def count_verb(text):
    return np.sum([1 if 'VB' in x[1] else 0 for x in text]) / max(len(text),1)

def count_adj(text):
    return np.sum([1 if 'JJ' in x[1] else 0 for x in text]) / max(len(text),1)

def count_det(text):
    return np.sum([1 if 'DT' in x[1] else 0 for x in text]) / max(len(text),1)

train_feats['profane'] = train_data.map(count_profane)
train_feats['capital'] = train_data.map(count_capital)
train_feats['punct'] = train_data.map(count_punct)
train_feats['pronoun'] = train_tags.map(count_pronoun)
train_feats['verb'] = train_tags.map(count_verb)
train_feats['adject'] = train_tags.map(count_adj)
train_feats['deter'] = train_tags.map(count_det)

test_feats['profane'] = test_data.fillna('nan').map(count_profane)
test_feats['capital'] = test_data.fillna('nan').map(count_capital)
test_feats['punct'] = test_data.fillna('nan').map(count_punct)
test_feats['pronoun'] = test_tags.map(count_pronoun)
test_feats['verb'] = test_tags.map(count_verb)
test_feats['adject'] = test_tags.map(count_adj)
test_feats['deter'] = test_tags.map(count_det)

train_feats.to_csv('../data/download/train_feats.csv', index=False)
test_feats.to_csv('../data/download/test_feats.csv', index=False)

### scale data

train_feats = pd.read_csv('../data/download/train_feats.csv')
test_feats = pd.read_csv('../data/download/test_feats.csv')

scaler = StandardScaler()
train_feats.iloc[:,1:] = scaler.fit_transform(train_feats.iloc[:,1:])
test_feats.iloc[:,1:] = scaler.transform(test_feats.iloc[:,1:])

train_feats.head(10)

test_feats.head(10)

train_feats.describe().round(4).T

test_feats.describe().round(4).T

train_feats.to_csv('../data/download/train_feats.csv', index=False)
test_feats.to_csv('../data/download/test_feats.csv', index=False)
