==> 04a_simple_stack.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import roc_curve, auc
pd.options.mode.chained_assignment = None

### datasets

train_score_1 = pd.read_csv('../data/model/baseline_1.csv')
train_score_1.columns = [train_score_1.columns[0]] + ['m1_' + x for x in train_score_1.columns[1:]]
train_score_2 = pd.read_csv('../data/model/baseline_2.csv')
train_score_2.columns = [train_score_2.columns[0]] + ['m2_' + x for x in train_score_2.columns[1:]]
train_score_3 = pd.read_csv('../data/model/baseline_3.csv')
train_score_3.columns = [train_score_2.columns[0]] + ['m3_' + x for x in train_score_3.columns[1:]]
train_labels = pd.read_csv('../data/download/train.csv').drop('comment_text', axis=1)
train_data = train_score_1.merge(train_score_2, on='id').merge(train_score_3, on='id')
train_labels = train_labels.sort_values(by='id').reset_index(drop=True)
train_ids = train_labels[['id']].copy()
train_data = train_data.drop('id', axis=1)
train_labels = train_labels.drop('id', axis=1)
print('train_data:', train_data.shape, train_labels.shape)
del train_score_1, train_score_2

test_score_1 = pd.read_csv('../data/submit/baseline_1.csv')
test_score_1.columns = [test_score_1.columns[0]] + ['m1_' + x for x in test_score_1.columns[1:]]
test_score_2 = pd.read_csv('../data/submit/baseline_2.csv')
test_score_2.columns = [test_score_2.columns[0]] + ['m2_' + x for x in test_score_2.columns[1:]]
test_score_3 = pd.read_csv('../data/submit/baseline_3.csv')
test_score_3.columns = [test_score_2.columns[0]] + ['m3_' + x for x in test_score_3.columns[1:]]
test_data = test_score_1.merge(test_score_2, on='id').merge(test_score_3, on='id')
test_ids = test_data[['id']].copy()
test_data = test_data.drop('id', axis=1)
print('test_data:', test_data.shape)
del test_score_1, test_score_2

### model

def eval_metric(labels, predict):
    fpr, tpr, threshold = roc_curve(labels, predict)
    return round(auc(fpr, tpr),4)
    
def model(label):
    feats = ['m1_' + label, 'm2_' + label, 'm3_' + label]
    scores = train_data[feats]
    scores[label] = scores.apply(lambda x : np.mean(x), axis=1)
    labels = train_labels[label]
    print('auc:', eval_metric(labels,scores[label]))
    scores = test_data[feats]
    scores[label] = scores.apply(lambda x : np.mean(x), axis=1)
    submit = test_ids.copy()
    submit[label] = scores[label].copy()
    return submit

toxic = model('toxic')
severe_toxic = model('severe_toxic')
obscene = model('obscene')
threat = model('threat')
insult = model('insult')
identity_hate = model('identity_hate')

submit = toxic.copy()
submit = submit.merge(severe_toxic, on='id')
submit = submit.merge(obscene, on='id')
submit = submit.merge(threat, on='id')
submit = submit.merge(insult, on='id')
submit = submit.merge(identity_hate, on='id')

submit.head()

submit.to_csv('../data/submit/baseline_4.csv', index=False)

(0.9821 + 0.9904 + 0.991 + 0.9916 + 0.9867 + 0.9854) / 6
